Reorganize data utils.
