+++ src/antsibull_nox/data/no-unwanted-files.py
@@ -12,7 +12,11 @@ from __future__ import annotations
 import os
 import sys
 
-from antsibull_nox.data_util import get_bool, get_list_of_strings, setup
+from antsibull_nox.data.antsibull_nox_data_util import (
+    get_bool,
+    get_list_of_strings,
+    setup,
+)
 
 
 def main() -> int:
